# openapi.storefront.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Commerce Platform Storefront API",
    "description": "Theme-facing page models and cart (spec §8.2). Reached through the edge only.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/storefront/v1/checkout": {
      "get": {
        "tags": [
          "storefront"
        ],
        "summary": "The checkout of the handed-off cart: lines and totals, the market's shipping methods with\nlive rates, payment methods, the selections so far and what is still missing.",
        "operationId": "get_checkout",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Cart capability from the `cart` cookie (shop) or `__Host-cart` (checkout origin).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutView"
                }
              }
            }
          },
          "404": {
            "description": "No checkout cart",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/storefront/v1/checkout/addresses": {
      "put": {
        "tags": [
          "storefront"
        ],
        "summary": "Billing and delivery address. The delivery country must be one the market ships to and\nthe tax profile covers (A3); VAT follows it.",
        "operationId": "put_addresses",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Cart capability from the `cart` cookie (shop) or `__Host-cart` (checkout origin).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressesInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutView"
                }
              }
            }
          },
          "422": {
            "description": "invalid_address | ship_to_not_allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/storefront/v1/checkout/contact": {
      "put": {
        "tags": [
          "storefront"
        ],
        "operationId": "put_contact",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Cart capability from the `cart` cookie (shop) or `__Host-cart` (checkout origin).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutView"
                }
              }
            }
          },
          "422": {
            "description": "invalid_email | invalid_phone",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/storefront/v1/checkout/fake-pay/{attempt}": {
      "get": {
        "tags": [
          "storefront"
        ],
        "summary": "What the fake pay page shows. `404` unless `PAYMENTS_FAKE=1`.",
        "operationId": "fake_pay_page",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FakePayment"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "storefront"
        ],
        "summary": "The fake pay page's Succeed/Fail: the fake provider signs an event and it is processed\nexactly like `POST /webhooks/fake` (signature, amount and currency checked first).",
        "operationId": "fake_pay",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FakePayInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FakePayment"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "attempt_finished",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/storefront/v1/checkout/payment": {
      "put": {
        "tags": [
          "storefront"
        ],
        "operationId": "put_payment",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Cart capability from the `cart` cookie (shop) or `__Host-cart` (checkout origin).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutView"
                }
              }
            }
          },
          "422": {
            "description": "unknown_payment_method | cod_not_allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/storefront/v1/checkout/place-order": {
      "post": {
        "tags": [
          "storefront"
        ],
        "summary": "Places the order (A12): one order per cart, `Idempotency-Key` required (a retry with the\nsame key answers the same order with `Idempotent-Replayed: true` and a fresh order token).\n`version` and `total_minor` are what the customer saw: `409 cart_changed` /\n`409 price_changed` when they no longer hold. Stock is reserved and the coupon redeemed in\nthe same transaction; then the payment attempt is initialized at the provider (A10).",
        "operationId": "place_order",
        "parameters": [
          {
            "name": "X-Storefront-Token",
            "in": "header",
            "description": "The tenant's public storefront token.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Market",
            "in": "header",
            "description": "Market id resolved from the shop host by the edge.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Locale",
            "in": "header",
            "description": "Locale hint (one of the market's locales).",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "X-Cart-Token",
            "in": "header",
            "description": "Cart capability from the `cart` cookie (shop) or `__Host-cart` (checkout origin).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "1-255 visible ASCII characters; kept for 24 hours.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaceOrderInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlacedOrder"
                }
              }
            }
          },
          "400": {
            "description": "idempotency_key_required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No checkout cart",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "order_already_placed | cart_changed | price_changed | insufficient_stock | out_of_stock | cart_unavailable | coupon_* | idempotency_conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "legal_consent_required | checkout_incomplete | ship_to_not_allowed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutView": {
        "type": "object",
        "description": "Everything the one-page checkout renders. Recomputed on every read.",
        "required": [
          "cart",
          "ship_to_countries",
          "shipping_methods",
          "payment_methods",
          "totals",
          "legal",
          "missing"
        ],
        "properties": {
          "billing_address": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CheckoutAddress"
              },
              {
                "type": "null"
              }
            ]
          },
          "cart": {
            "$ref": "#/components/schemas/CartView"
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "legal": {
            "$ref": "#/components/schemas/Legal"
          },
          "missing": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "What is still missing before the order can be placed (`email`, `billing_address`,\n`shipping_method`, `pickup_point`, `payment_method`, `cart_unavailable`)."
          },
          "packeta": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PacketaWidget"
              },
              {
                "type": "null"
              }
            ]
          },
          "payment_method": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/MethodKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "payment_methods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentOption"
            }
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "pickup_point": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PickupPoint"
              },
              {
                "type": "null"
              }
            ]
          },
          "ship_to_countries": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "A3: where this market delivers."
          },
          "shipping_address": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/CheckoutAddress"
              },
              {
                "type": "null"
              }
            ]
          },
          "shipping_method_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "shipping_methods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShippingOption"
            }
          },
          "totals": {
            "$ref": "#/components/schemas/Totals"
          }
        }
      },
...
      "FakePayInput": {
        "type": "object",
        "required": [
          "outcome"
        ],
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/Outcome"
          }
        },
        "additionalProperties": false
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details body.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code, e.g. `not_found`."
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`; `code` carries the specific error kind."
          }
        }
      },
...
    }
  },
...
}